// www/js/Transactions/transactions/ts/enums/TransactionItemType.ts
enum TransactionItemType {
  AffiliatePayout = 'AffiliatePayout',
  Asset = 'Asset',
  Bundle = 'Bundle',
  DeveloperProduct = 'DeveloperProduct',
  GamePass = 'GamePass',
  Place = 'Place',
  PrivateServer = 'PrivateServer',
  RobloxProduct = 'RobloxProduct',
  SubscriptionProduct = 'SubscriptionProduct',
  RobloxSelectTransfer = 'RobloxSelectTransfer',
  CreatorRewardsPayout = 'CreatorRewardsPayout'
}

export default TransactionItemType;